refactor: move bulit-in category component
